refactor(dashboard): move _to_json_safe to common projection layer
